@@ -1,16 +1,13 @@
 package com.couchbase.travelsample.bookings;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.couchbase.lite.Array;
-import com.couchbase.lite.ArrayExpression;
 import com.couchbase.lite.CouchbaseLiteException;
 import com.couchbase.lite.DataSource;
 import com.couchbase.lite.Database;
 import com.couchbase.lite.Expression;
-import com.couchbase.lite.Dictionary;
 import com.couchbase.lite.MutableDictionary;
-import com.couchbase.lite.MutableDocument;
 import com.couchbase.lite.Query;
 import com.couchbase.lite.QueryBuilder;
 import com.couchbase.lite.QueryChange;
