@@ -2,9 +2,9 @@ package com.couchbase.travelsample.searchflight;
 
 import android.os.Build;
 import android.os.Handler;
-import android.support.annotation.MainThread;
-import android.support.annotation.NonNull;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.MainThread;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import android.util.Log;
 
 import com.couchbase.lite.CouchbaseLiteException;
